rename: 책임 변경에 따른 이름 변경 - Racer 클래스를 InputValidator로 변경 - 기존 Racer의 자동차 경주에 대한 책임이 Cars로 넘어가면서 입력값 검증 로직만 남음
