+++ src/main/java/org/duckstudy/GameApplication.java
@@ -10,7 +10,7 @@ import org.duckstudy.generator.Generator;
 import org.duckstudy.generator.RandomValueGenerator;
 import org.duckstudy.movingcar.Car;
 import org.duckstudy.movingcar.Cars;
-import org.duckstudy.racer.Racer;
+import org.duckstudy.validator.InputValidator;
 
 public class GameApplication {
     public static void main(String[] args) throws IOException {
@@ -20,7 +20,7 @@ public class GameApplication {
 
         Generator generator = new RandomValueGenerator();
         Cars cars = new Cars(nameList.length, nameList, generator);
-        makeCarRacer(nameList, repetitionNum, cars);
+        validateInput(nameList, repetitionNum, cars);
         printGameResult(cars, repetitionNum);
     }
 
@@ -34,10 +34,10 @@ public class GameApplication {
         return Integer.parseInt(br.readLine());
     }
 
-    private static Optional<Racer> makeCarRacer(String[] nameList, int repetitionNum, Cars cars) {
+    private static Optional<InputValidator> validateInput(String[] nameList, int repetitionNum, Cars cars) {
         try {
-            Racer racer = new Racer(repetitionNum, nameList);
-            return Optional.of(racer);
+            InputValidator inputValidator = new InputValidator(repetitionNum, nameList);
+            return Optional.of(inputValidator);
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             return Optional.empty();
